backend/requirements.txt
# FastAPI e dependências
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Gemini
google-generativeai==0.3.2

# Monitoramento
prometheus-client==0.19.0

# Utilitários
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0
